@@ -27,18 +27,23 @@
         ]
     },
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "guzzlehttp/guzzle": "^7.0",
         "aws/aws-sdk-php": "~3.0",
-        "illuminate/contracts": "^9.0|^10.0|^11.0",
-        "illuminate/database": "^9.0|^10.0|^11.0",
-        "illuminate/queue": "^9.0|^10.0|^11.0",
-        "illuminate/console": "^9.0|^10.0|^11.0",
+        "illuminate/contracts": "^10.0|^11.0|^12.0",
+        "illuminate/database": "^10.0|^11.0|^12.0",
+        "illuminate/queue": "^10.0|^11.0|^12.0",
+        "illuminate/console": "^10.0|^11.0|^12.0",
         "twig/twig": "^3.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^10.0|^11.0",
         "mockery/mockery": "^1.5"
+    },
+    "config": {
+        "platform": {
+            "php": "8.2"
+        }
     },
     "extra": {
         "laravel": {
